fix array ref
